% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_11_kRp.corpus-class_clozeDelete.R
\docType{methods}
\name{clozeDelete,kRp.corpus-method}
\alias{clozeDelete,kRp.corpus-method}
\title{Apply clozeDelete() to all texts in kRp.corpus objects}
\usage{
\S4method{clozeDelete}{kRp.corpus}(obj, mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:clozeDelete]{clozeDelete}}.}
}
\value{
An object of the same class as \code{obj}.
}
\description{
This method calls \code{\link[koRpus:clozeDelete]{clozeDelete}} on all tagged text objects
inside the given \code{obj} object (using \code{mclapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
# remove all punctuation
myCorpus <- clozeDelete(myCorpus)
}
}
