% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_15_kRp.corpus-class_corpus2tagged.R
\docType{methods}
\name{corpus2tagged}
\alias{corpus2tagged}
\alias{corpus2tagged,kRp.corpus-method}
\alias{corpus2tagged,-methods}
\title{Turn a kRp.corpus object into a list of kRp.tagged objects}
\usage{
corpus2tagged(obj)

\S4method{corpus2tagged}{kRp.corpus}(obj)
}
\arguments{
\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}}.}
}
\value{
A named list of objects of class \code{\link[koRpus:kRp.tagged-class]{kRp.tagged}}.
   Elements are named by their \code{doc_id}.
}
\description{
For some analysis steps it might be important to have individual tagged texts
instead of one large corpus object. This method produces just that.
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(path.package("tm.plugin.koRpus"), "tests", "testthat", "samples"),
  hierarchy=list(
    Topic=c(
      C3S="C3S",
      GEMA="GEMA"
    ),
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)

myCorpusList <- corpus2tagged(myCorpus)
}
}
