% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_12_kRp.corpus-class_cTest.R
\docType{methods}
\name{cTest,kRp.hierarchy-method}
\alias{cTest,kRp.hierarchy-method}
\title{Apply cTest() to all texts in kRp.hierarchy objects}
\usage{
\S4method{cTest}{kRp.hierarchy}(obj, mc.cores = getOption("mc.cores",
  1L), ...)
}
\arguments{
\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:cTest]{cTest}}.}
}
\value{
An object of the same class as \code{obj}.
}
\description{
This method calls \code{\link[koRpus:cTest]{cTest}} on all tagged text objects
inside the given \code{obj} object (using \code{lapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
# remove all punctuation
myCorpus <- cTest(myCorpus)
}
}
