% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{simpleCorpus}
\alias{simpleCorpus}
\alias{sourcesCorpus}
\alias{topicCorpus}
\title{Deprecated functions and methods}
\usage{
simpleCorpus(dir = ".", lang = "kRp.env", tagger = "kRp.env",
  encoding = "", pattern = NULL, recursive = FALSE,
  ignore.case = FALSE, mode = "text", source = "", topic = "",
  format = "file", mc.cores = getOption("mc.cores", 1L), ...)

sourcesCorpus(path, sources, topic = "", format = "file",
  mc.cores = getOption("mc.cores", 1L), ...)

topicCorpus(paths, sources, format = "file",
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{dir}{Use \code{readCorpus} instead.}

\item{lang}{Use \code{readCorpus} instead.}

\item{tagger}{Use \code{readCorpus} instead.}

\item{encoding}{Use \code{readCorpus} instead.}

\item{pattern}{Use \code{readCorpus} instead.}

\item{recursive}{Use \code{readCorpus} instead.}

\item{ignore.case}{Use \code{readCorpus} instead.}

\item{mode}{Use \code{readCorpus} instead.}

\item{source}{Use \code{readCorpus} instead.}

\item{topic}{Use \code{readCorpus} instead.}

\item{format}{Use \code{readCorpus} instead.}

\item{mc.cores}{Use \code{readCorpus} instead.}

\item{...}{Use \code{readCorpus} instead.}

\item{path}{Use \code{readCorpus} instead.}

\item{sources}{Use \code{readCorpus} instead.}

\item{paths}{Use \code{readCorpus} instead.}
}
\description{
These functions were used in earlier versions of the package but since
replaced by \code{\link[tm.plugin.koRpus:readCorpus]{readCorpus}}.
}
