% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_14_kRp.corpus-class_docTermMatrix.R
\docType{methods}
\name{docTermMatrix}
\alias{docTermMatrix}
\alias{docTermMatrix,kRp.hierarchy-method}
\alias{docTermMatrix,-methods}
\title{Generate a document-term matrix from a corpus opbject}
\usage{
docTermMatrix(obj, terms = "token", case.sens = FALSE, ...)

\S4method{docTermMatrix}{kRp.hierarchy}(obj, terms = "token",
  case.sens = FALSE)
}
\arguments{
\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{terms}{A character string defining the \code{TT.res} column to be used for calculating the matrix.}

\item{case.sens}{Logical, whether terms should be counted case sensitive.}
}
\value{
A sparse matrix of class \code{\link[Matrix:dgCMatrix]{dgCMatrix}}.
}
\description{
Generate a document-term matrix from a corpus opbject
}
