% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_09_kRp.corpus-class_filterByClass.R
\docType{methods}
\name{filterByClass,kRp.corpus-method}
\alias{filterByClass,kRp.corpus-method}
\title{Apply filterByClass() to all texts in kRp.corpus objects}
\usage{
\S4method{filterByClass}{kRp.corpus}(txt, mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{txt}{An object of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:filterByClass]{filterByClass}}.}
}
\value{
An object of the same class as \code{txt}.
}
\description{
This method calls \code{\link[koRpus:filterByClass]{filterByClass}} on all tagged text objects
inside the given \code{txt} object (using \code{mclapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "examples", "corpus", "Winner"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_prev="Wikipedia (old)",
      Wikipedia_new="Wikipedia (new)"
    )
  )
)
# remove all punctuation
myCorpus <- filterByClass(myCorpus)
}
}
