% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02_method_02_kRp.corpus-class_hyphen.corpus.R
\docType{methods}
\name{hyphen.corpus}
\alias{hyphen.corpus}
\alias{hyphen.corpus,kRp.corpus-method}
\title{Apply hyphen() to all texts in kRp.corpus objects}
\usage{
hyphen.corpus(corpus, ...)

\S4method{hyphen.corpus}{kRp.corpus}(corpus, ...)
}
\arguments{
\item{corpus}{An object of class \code{\link[tm.plugin.koRpus]{kRp.corpus-class}}.}

\item{...}{options to pass through to \code{\link[koRpus:hyphen]{hyphen}}.}
}
\description{
This method calls \code{\link[koRpus:hyphen]{hyphen}} on all tagged text objects
inside the given \code{corpus} object (using \code{lapply}).
}
\examples{
\dontrun{
myTexts <- kRp.Corpus(dir=file.path("/home","me","textCorpus"))
myTexts <- hyphen.corpus(myTexts)
}
}

