% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_02_kRp.corpus-class_hyphen.R
\docType{methods}
\name{hyphen,kRp.hierarchy-method}
\alias{hyphen,kRp.hierarchy-method}
\title{Apply hyphen() to all texts in kRp.hierarchy objects}
\usage{
\S4method{hyphen}{kRp.hierarchy}(words, mc.cores = getOption("mc.cores", 1L),
  quiet = TRUE, ...)
}
\arguments{
\item{words}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{quiet}{Logical,
      if \code{FALSE} shows a status bar for the hyphenation process of each text.}

\item{...}{options to pass through to \code{\link[sylly:hyphen]{hyphen}}.}
}
\value{
An object of the same class as \code{words}.
}
\description{
This method calls \code{\link[sylly:hyphen]{hyphen}} on all tagged text objects
inside the given \code{words} object (using \code{lapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S",
      "Wikipedia_alt"
  )
)
myCorpus <- hyphen(myCorpus)
}
}
