% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01_class_01_kRp.corpus.R
\docType{class}
\name{kRp.corpus,-class}
\alias{kRp.corpus,-class}
\alias{kRp.corpus-class}
\title{S4 Class kRp.corpus}
\description{
Objects of this class can contain multiple texts simultaneously. It supports both the \code{tm} package's
\code{\link[tm]{Corpus}} class and \code{koRpus}' own object classes and stores them in separated slots.
}
\details{
Objects can be created using the \code{\link[tm.plugin.koRpus:kRp.Corpus]{kRp.Corpus}} function.
}
\section{Slots}{

\describe{
\item{\code{meta}}{A named list. Can be used to store meta information. Currently,
      no particular format is defined.}

\item{\code{raw}}{A list of objects of class \code{\link[tm]{Corpus}}.}

\item{\code{tagged}}{A list of objects of class \code{\link[koRpus]{kRp.taggedText-class}}.}

\item{\code{hyphen}}{A list of objects of class \code{\link[koRpus]{kRp.hyphen-class}}.}

\item{\code{TTR}}{A list of objects of class \code{\link[koRpus]{kRp.TTR-class}}.}

\item{\code{readability}}{A list of objects of class \code{\link[koRpus]{kRp.readability-class}}.}
}}
\note{
There is also methods to transform get the lists of all particular slots from objects of this class.
}
\keyword{classes}

