% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02_method_20_kRp.corpus_get_set_is.R
\docType{methods}
\name{corpusTagged}
\alias{corpusFreq}
\alias{corpusFreq,-methods}
\alias{corpusFreq,kRp.corpus-method}
\alias{corpusFreq,kRp.sourcesCorpus-method}
\alias{corpusFreq,kRp.topicCorpus-method}
\alias{corpusFreq<-}
\alias{corpusFreq<-,-methods}
\alias{corpusFreq<-,kRp.corpus-method}
\alias{corpusFreq<-,kRp.sourcesCorpus-method}
\alias{corpusFreq<-,kRp.topicCorpus-method}
\alias{corpusHyphen}
\alias{corpusHyphen,-methods}
\alias{corpusHyphen,kRp.corpus-method}
\alias{corpusHyphen<-}
\alias{corpusHyphen<-,-methods}
\alias{corpusHyphen<-,kRp.corpus-method}
\alias{corpusMeta}
\alias{corpusMeta,-methods}
\alias{corpusMeta,kRp.corpus-method}
\alias{corpusMeta<-}
\alias{corpusMeta<-,-methods}
\alias{corpusMeta<-,kRp.corpus-method}
\alias{corpusReadability}
\alias{corpusReadability,-methods}
\alias{corpusReadability,kRp.corpus-method}
\alias{corpusReadability<-}
\alias{corpusReadability<-,-methods}
\alias{corpusReadability<-,kRp.corpus-method}
\alias{corpusSources}
\alias{corpusSources,-methods}
\alias{corpusSources,kRp.sourcesCorpus-method}
\alias{corpusSources<-}
\alias{corpusSources<-,-methods}
\alias{corpusSources<-,kRp.sourcesCorpus-method}
\alias{corpusTTR}
\alias{corpusTTR,-methods}
\alias{corpusTTR,kRp.corpus-method}
\alias{corpusTTR<-}
\alias{corpusTTR<-,-methods}
\alias{corpusTTR<-,kRp.corpus-method}
\alias{corpusTagged}
\alias{corpusTagged,-methods}
\alias{corpusTagged,kRp.corpus-method}
\alias{corpusTagged<-}
\alias{corpusTagged<-,-methods}
\alias{corpusTagged<-,kRp.corpus-method}
\alias{corpusTm}
\alias{corpusTm,-methods}
\alias{corpusTm,kRp.corpus-method}
\alias{corpusTm<-}
\alias{corpusTm<-,-methods}
\alias{corpusTm<-,kRp.corpus-method}
\alias{corpusTopics}
\alias{corpusTopics,-methods}
\alias{corpusTopics,kRp.topicCorpus-method}
\alias{corpusTopics<-}
\alias{corpusTopics<-,-methods}
\alias{corpusTopics<-,kRp.topicCorpus-method}
\alias{is.corpus}
\title{Getter/setter methods for kRp.corpus objects}
\usage{
corpusTagged(obj)

\S4method{corpusTagged}{kRp.corpus}(obj)

corpusTagged(obj) <- value

\S4method{corpusTagged}{kRp.corpus}(obj) <- value

corpusReadability(obj)

\S4method{corpusReadability}{kRp.corpus}(obj)

corpusReadability(obj) <- value

\S4method{corpusReadability}{kRp.corpus}(obj) <- value

corpusTm(obj)

\S4method{corpusTm}{kRp.corpus}(obj)

corpusTm(obj) <- value

\S4method{corpusTm}{kRp.corpus}(obj) <- value

corpusMeta(obj, meta = NULL, fail = TRUE)

\S4method{corpusMeta}{kRp.corpus}(obj, meta = NULL, fail = TRUE)

corpusMeta(obj, meta = NULL) <- value

\S4method{corpusMeta}{kRp.corpus}(obj, meta = NULL) <- value

corpusHyphen(obj)

\S4method{corpusHyphen}{kRp.corpus}(obj)

corpusHyphen(obj) <- value

\S4method{corpusHyphen}{kRp.corpus}(obj) <- value

corpusTTR(obj)

\S4method{corpusTTR}{kRp.corpus}(obj)

corpusTTR(obj) <- value

\S4method{corpusTTR}{kRp.corpus}(obj) <- value

corpusSources(obj, source = NULL)

\S4method{corpusSources}{kRp.sourcesCorpus}(obj, source = NULL)

corpusSources(obj, source = NULL) <- value

\S4method{corpusSources}{kRp.sourcesCorpus}(obj, source = NULL) <- value

corpusTopics(obj, topic = NULL)

\S4method{corpusTopics}{kRp.topicCorpus}(obj, topic = NULL)

corpusTopics(obj, topic = NULL) <- value

\S4method{corpusTopics}{kRp.topicCorpus}(obj, topic = NULL) <- value

corpusFreq(obj)

\S4method{corpusFreq}{kRp.corpus}(obj)

\S4method{corpusFreq}{kRp.sourcesCorpus}(obj)

\S4method{corpusFreq}{kRp.topicCorpus}(obj)

corpusFreq(obj, topic) <- value

\S4method{corpusFreq}{kRp.corpus}(obj) <- value

\S4method{corpusFreq}{kRp.sourcesCorpus}(obj) <- value

\S4method{corpusFreq}{kRp.topicCorpus}(obj) <- value

is.corpus(obj)
}
\arguments{
\item{obj}{An object of class \code{\link[tm.plugin.koRpus]{kRp.corpus-class}},
\code{\link[tm.plugin.koRpus]{kRp.sourcesCorpus-class}} or
\code{\link[tm.plugin.koRpus]{kRp.topicCorpus-class}}.}

\item{value}{The new value to replace the current with.}

\item{meta}{If not NULL, the \code{meta} list entry of the given name.}

\item{fail}{Logical,
      whether the method should fail with an error if \code{meta} was not found.
If set to \code{FALSE}, returns \code{invisible(NULL)} instead.}

\item{source}{Character value, name of a source in this object. If given,
      you get/set only the value of this source.}

\item{topic}{Character value, name of a topic in this object. If given,
      you get/set only the value of this topic.}
}
\description{
These methods should be used to get or set values of text objects
generated by functions like \code{\link[tm.plugin.koRpus:simpleCorpus]{simpleCorpus}}.
}
\details{
\itemize{
  \item{\code{corpusTagged()} }{returns the list of \code{kRp.tagged} objects.}
  \item{\code{corpusReadability()} }{returns the list of \code{kRp.readability} objects.}
  \item{\code{corpusTm()} }{returns the \code{VCorpus} object.}
  \item{\code{corpusMeta()} }{returns the list with meta information.}
  \item{\code{corpusHyphen()} }{returns the list of \code{kRp.hyphen} objects.}
  \item{\code{corpusTTR()} }{returns the list of \code{kRp.TTR} objects.}
}
}
\examples{
\dontrun{
corpusTagged(myCorpus)
corpusMeta(myCorpus, "note") <- "an interesting read!"
}
}

