% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_01_kRp.flatHier.R
\docType{class}
\name{kRp.flatHier,-class}
\alias{kRp.flatHier,-class}
\alias{kRp_flatHier}
\alias{kRp.flatHier-class}
\title{S4 Class kRp.flatHier}
\description{
Objects of this class can contain full text corpora in a hierachical structure. It supports both the \code{tm} package's
\code{\link[tm]{Corpus}} class and \code{koRpus}' own object classes and stores them in separated slots.
}
\details{
Objects should be created using the \code{\link[tm.plugin.koRpus:readCorpus]{readCorpus}} function.
}
\section{Slots}{

\describe{
\item{\code{lang}}{A character string,
      naming the language that is assumed for the tokenized texts in this object.}

\item{\code{desc}}{A named list of descriptive statistics of the tagged texts.}

\item{\code{hierarchy}}{A named list of named character vectors describing the directory hierarchy level by level.}

\item{\code{meta}}{A named list. Can be used to store meta information. Currently,
      no particular format is defined.}

\item{\code{raw}}{A list of objects of class \code{\link[tm]{Corpus}}.}

\item{\code{TT.res}}{A data frame as used for the \code{TT.res} slot in objects of class \code{kRp.taggedText}. In addition to the columns
usually found in those objects,
      this data frame also has a factor column for each hierarchical category defined (if any).}

\item{\code{summary}}{A summary data frame for the full corpus.}

\item{\code{hyphen}}{A named list of objects of class \code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}.}

\item{\code{readability}}{A named list of objects of class \code{\link[koRpus:kRp.readability-class]{kRp.readability}}.}

\item{\code{freq.analysis}}{A list with information on the word frequencies of the analyzed text.}

\item{\code{freq}}{A list with two elements,
      \code{freq.analysis} and \code{corpus}. \code{freq.analysis} contains the
\code{freq.analysis} slot of a \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}} class object after
\code{\link[tm.plugin.koRpus:freq.analysis]{freq.analysis}} was called,
      whereas \code{corpus} holds the results of a call to
\code{\link[tm.plugin.koRpus:read.corp.custom]{read.corp.custom}}.}

\item{\code{diff}}{A named list of \code{diff} slots of a \code{\link[sylly:kRp.text.trans-class]{kRp.text.trans}} object after
a method like \code{\link[tm.plugin.koRpus:textTransform]{textTransform}} was called.}
}}

\note{
There is also \code{\link[tm.plugin.koRpus:kRp.flatHier_get-methods]{getter and setter methods}} for objects of this class.
}
\section{Contructor function}{

Should you need to manually generate objects of this class (which should rarely be the case),
      the contructor function 
\code{kRp_flatHier(...)} can be used instead of
\code{new("kRp.flatHier", ...)}. Whenever possible, stick to
\code{\link[tm.plugin.koRpus:readCorpus]{readCorpus}}.
}

\examples{
\dontrun{
# use readCorpus() to create objects of class kRp.flatHier
myCorpus <- readCorpus(
  dir=file.path(path.package("tm.plugin.koRpus"), "tests", "testthat", "samples"),
  hierarchy=list(
    Topic=c(
      C3S="C3S",
      GEMA="GEMA"
    ),
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
}
# manual creation
emptyCorpus <- kRp_hierarchy()
}
\keyword{classes}
