% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_01_kRp.hierarchy.R
\docType{class}
\name{kRp.hierarchy,-class}
\alias{kRp.hierarchy,-class}
\alias{kRp_hierarchy}
\alias{kRp.hierarchy-class}
\title{S4 Class kRp.hierarchy}
\description{
Objects of this class can contain full text corpora in a hierachical structure. It supports both the \code{tm} package's
\code{\link[tm]{Corpus}} class and \code{koRpus}' own object classes and stores them in separated slots.
}
\details{
Objects should be created using the \code{\link[tm.plugin.koRpus:readCorpus]{readCorpus}} function.
}
\section{Slots}{

\describe{
\item{\code{level}}{A numeric value defining the hierachical level. Objects of this class can be nested,
      \code{level=0} is the
deepest (i.e., a collection of actual texts),
      and higher values indicate that the object represents only a category.}

\item{\code{category}}{A character string describing the category of this level (e.g.,
      "source" or "topic").}

\item{\code{id}}{A character string naming this category level (i.e.,
      a particular source or topic).}

\item{\code{path}}{A character string,
      full path to the directory of the current category level.}

\item{\code{files}}{A list of character strings with only the file names of all texts.}

\item{\code{children}}{If \code{level > 0} a list of objects of class \code{kRp.hierarchy},
      otherwise an empty list.}

\item{\code{summary}}{A summary data frame for the full corpus at the current level.}

\item{\code{meta}}{A named list. Can be used to store meta information. Currently,
      no particular format is defined.}

\item{\code{raw}}{A list of objects of class \code{\link[tm]{Corpus}}. Only used at \code{level=0}.}

\item{\code{tagged}}{A list of objects of class \code{kRp.taggedText} (a class union for tagged text objects). Only used at \code{level=0}.}

\item{\code{hyphen}}{A list of objects of class \code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}. Only used at \code{level=0}.}

\item{\code{TTR}}{A list of objects of class \code{\link[koRpus:kRp.TTR-class]{kRp.TTR}}. Only used at \code{level=0}.}

\item{\code{readability}}{A list of objects of class \code{\link[koRpus:kRp.readability-class]{kRp.readability}}. Only used at \code{level=0}.}

\item{\code{freq}}{A list with two elements,
      \code{texts} and \code{corpus}. At \code{level=0}, both hold objects of class \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}},
where \code{texts} is a list of these objects (one for each text),
      and \code{corpus} is a single object for the full corpus. At higher levels only \code{corpus}
is used.}
}}

\note{
There is also \code{\link[tm.plugin.koRpus:kRp.hierarchy_get-methods]{getter and setter methods}} for objects of this class.
}
\section{Contructor function}{

Should you need to manually generate objects of this class (which should rarely be the case),
      the contructor function 
\code{kRp_hierarchy(...)} can be used instead of
\code{new("kRp.hierarchy", ...)}. Whenever possible, stick to
\code{\link[tm.plugin.koRpus:readCorpus]{readCorpus}}.
}

\examples{
\dontrun{
# use readCorpus() to create objects of class kRp.hierarchy
myCorpus <- readCorpus(
  dir=file.path(path.package("tm.plugin.koRpus"), "tests", "testthat", "samples"),
  hierarchy=list(
    Topic=c(
      C3S="C3S",
      GEMA="GEMA"
    ),
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
}
# manual creation
emptyCorpus <- kRp_hierarchy()
}
\keyword{classes}
