% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_03_kRp.topicCorpus.R
\docType{class}
\name{kRp.topicCorpus,-class}
\alias{kRp.topicCorpus,-class}
\alias{kRp_topicCorpus}
\alias{kRp.topicCorpus-class}
\title{S4 Class kRp.topicCorpus and contructor}
\description{
Objects of this class can contain multiple texts simultaneously. Adding to that,
      these texts can be ordered
at two levels: topic and source. This is useful for comparisons if you have a defined number of topics and
textst from different sources on these topics. Note however,
      that there must be at least one text on each topic
in all of the given sources.
}
\details{
For each combination of topic and source,
      there is exactly one object of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}}.

Objects should be created using the \code{\link[tm.plugin.koRpus:topicCorpus]{topicCorpus}} function.
}
\section{Slots}{

\describe{
\item{\code{summary}}{A summary data.frame for all topics combined.}

\item{\code{topics}}{A named list of nested objects. Each element is named after a topic and contains an object of class kRp.sourcesCorpus.}

\item{\code{freq}}{An object of class \code{\link[koRpus]{kRp.corp.freq-class}},
      can contain word frequency
information on the full corpus if this object was analysed with
\code{\link[tm.plugin.koRpus:read.corp.custom]{read.corp.custom}}.}
}}

\note{
There is also \code{\link[tm.plugin.koRpus:kRp.corpus_get-methods]{getter and setter methods}} for objects of this class.
}
\section{Contructor function}{

Should you need to manually generate objects of this class (which should rarely be the case),
      the contructor function 
\code{kRp_topicCorpus(...)} can be used instead of
\code{new("kRp.topicCorpus", ...)}. Whenever possible, stick to
\code{\link[tm.plugin.koRpus:topicCorpus]{topicCorpus}}.
}

\keyword{classes}
