% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kRpSource.R
\name{kRpSource}
\alias{kRpSource}
\title{A source function for tm}
\usage{
kRpSource(obj, encoding = "UTF-8")
}
\arguments{
\item{obj}{An object of class \code{\link[koRpus:kRp.text-class]{kRp.text}} (a class union for tagged text objects).}

\item{encoding}{Character string, defining the character encoding of the object.}
}
\value{
An object of class \code{\link[tm:Source]{Source}},
      also inheriting class \code{kRpSource}.
}
\description{
An rather untested attempt to sketch a \code{\link[tm:Source]{Source}} function for \code{tm}.
Supposed to be used to translate tagged \code{koRpus} objects into \code{tm} objects.
}
\details{
Also provided are the methods \code{getElem} and \code{pGetElem} for S3 class \code{kRpSource}.
}
