% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_03_kRp.corpus-class_lex.div.R
\docType{methods}
\name{lex.div,kRp.corpus-method}
\alias{lex.div,kRp.corpus-method}
\alias{lex.div,kRp.sourcesCorpus-method}
\alias{lex.div,kRp.topicCorpus-method}
\title{Apply lex.div() to all texts in kRp.corpus objects}
\usage{
\S4method{lex.div}{kRp.corpus}(txt, summary = TRUE, ...)

\S4method{lex.div}{kRp.sourcesCorpus}(txt, summary = TRUE, ...)

\S4method{lex.div}{kRp.topicCorpus}(txt, summary = TRUE, ...)
}
\arguments{
\item{txt}{An object of class \code{\link[tm.plugin.koRpus]{kRp.corpus-class}},
\code{\link[tm.plugin.koRpus]{kRp.sourcesCorpus-class}} or
\code{\link[tm.plugin.koRpus]{kRp.topicCorpus-class}}.}

\item{summary}{Logical, determines if the \code{summary} slot should automatically be
updated by calling \code{\link[tm.plugin.koRpus:summary]{summary}} on the result.}

\item{...}{options to pass through to \code{\link[koRpus:lex.div]{lex.div}}.}
}
\value{
An object of the same class as \code{txt}.
}
\description{
This method calls \code{\link[koRpus:lex.div]{lex.div}} on all tagged text objects
inside the given \code{txt} object (using \code{lapply}).
}
\examples{
\dontrun{
myTexts <- simpleCorpus(dir=file.path("/home","me","textCorpus"))
myTexts <- lex.div(myTexts)
}
}

