% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02_method_03_kRp.corpus-class_lex.div.corpus.R
\docType{methods}
\name{lex.div.corpus}
\alias{lex.div.corpus}
\alias{lex.div.corpus,kRp.corpus-method}
\title{Apply lex.div() to all texts in kRp.corpus objects}
\usage{
lex.div.corpus(corpus, ...)

\S4method{lex.div.corpus}{kRp.corpus}(corpus, ...)
}
\arguments{
\item{corpus}{An object of class \code{\link[tm.plugin.koRpus]{kRp.corpus-class}}.}

\item{...}{options to pass through to \code{\link[koRpus:lex.div]{lex.div}}.}
}
\description{
This method calls \code{\link[koRpus:lex.div]{lex.div}} on all tagged text objects
inside the given \code{corpus} object (using \code{lapply}).
}
\examples{
\dontrun{
myTexts <- kRp.Corpus(dir=file.path("/home","me","textCorpus"))
myTexts <- lex.div.corpus(myTexts)
}
}

