% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02_method_04_kRp.corpus-class_read.corp.custom.R
\docType{methods}
\name{read.corp.custom,kRp.corpus-method}
\alias{read.corp.custom}
\alias{read.corp.custom,kRp.corpus-method}
\alias{read.corp.custom,kRp.sourcesCorpus-method}
\alias{read.corp.custom,kRp.topicCorpus-method}
\title{Apply read.corp.custom() to all texts in kRp.corpus objects}
\usage{
\S4method{read.corp.custom}{kRp.corpus}(corpus, ...)

\S4method{read.corp.custom}{kRp.sourcesCorpus}(corpus, ...)

\S4method{read.corp.custom}{kRp.topicCorpus}(corpus, ...)
}
\arguments{
\item{corpus}{An object of class \code{\link[tm.plugin.koRpus]{kRp.corpus-class}},
\code{\link[tm.plugin.koRpus]{kRp.sourcesCorpus-class}} or
\code{\link[tm.plugin.koRpus]{kRp.topicCorpus-class}}.}

\item{...}{options to pass through to \code{\link[koRpus:read.corp.custom]{read.corp.custom}}.}
}
\description{
This method calls \code{\link[koRpus:read.corp.custom]{read.corp.custom}} on all tagged text objects
inside the given \code{corpus} object (using \code{lapply}).
}
\examples{
\dontrun{
myBasePath <- file.path("/home","me","textCorpus")
# analyse a single corpus
myTexts <- simpleCorpus(dir=myBasePath)
myTexts <- read.corp.custom(myTexts)

# you can also analyse multiple corpora from different sources simultaneously
# the following would assume that below 'myBasePath', there are subfolders
# named "Wikipedia" and Journal of Applied Geschwurbel"
mySources <- c(
   wp="Wikipedia",
   jg="Journal of Applied Geschwurbel"
 )
mySourcesTexts <- sourcesCorpus(
  path=myBasePath,
  sources=mySources
)
# this will also call read.corp.custom() recursively on
# the single text level
mySourcesTexts <- read.corp.custom(mySourcesTexts)

# and you might have guessed, you can also add a topic level
# with the following two vectors you would describe your
# data structure as two subfolders called "proc" and "gesw"
# below myBasePath, and two further subfolders named "Wikipedia"
# and Journal of Applied Geschwurbel" below each of the topic folders,
# containing the actual texts
myTopicPaths <- c(
  procrastination=file.path(myBasePath, "proc"),
  geschwurbel=file.path(myBasePath, "gesw")
)
mySources <- c(
   wp="Wikipedia",
   jg="Journal of Applied Geschwurbel"
 )
myTopicTexts <- topicCorpus(
  paths=myTopicPaths,
  sources=mySources
)
# this will also call read.corp.custom() recursively on
# the source and single text level
myTopicTexts <- read.corp.custom(myTopicTexts)

}
}

