% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02_method_01_kRp.corpus-class_readability.corpus.R
\docType{methods}
\name{readability.corpus}
\alias{readability.corpus}
\alias{readability.corpus,kRp.corpus-method}
\title{Apply readability() to all texts in kRp.corpus objects}
\usage{
readability.corpus(corpus, ...)

\S4method{readability.corpus}{kRp.corpus}(corpus, ...)
}
\arguments{
\item{corpus}{An object of class \code{\link[tm.plugin.koRpus]{kRp.corpus-class}}.}

\item{...}{options to pass through to \code{\link[koRpus:readability]{readability}}.}
}
\description{
This method calls \code{\link[koRpus:readability]{readability}} on all tagged text objects
inside the given \code{corpus} object (using \code{lapply}).
}
\examples{
\dontrun{
myTexts <- kRp.Corpus(dir=file.path("/home","me","textCorpus"))
myTexts <- readability.corpus(myTexts)
}
}

