% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_01_kRp.corpus-class_readability.R
\docType{methods}
\name{readability,kRp.hierarchy-method}
\alias{readability,kRp.hierarchy-method}
\title{Apply readability() to all texts in kRp.hierarchy objects}
\usage{
\S4method{readability}{kRp.hierarchy}(txt.file, summary = TRUE,
  mc.cores = getOption("mc.cores", 1L), quiet = TRUE, ...)
}
\arguments{
\item{txt.file}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{summary}{Logical, determines if the \code{summary} slot should automatically be
updated by calling \code{\link[tm.plugin.koRpus:summary]{summary}} on the result.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{quiet}{Logical,
      if \code{FALSE} shows a status bar for some calculations of each text, see
\code{\link[koRpus:readability]{readability}} for details.}

\item{...}{options to pass through to \code{\link[koRpus:readability]{readability}}.}
}
\value{
An object of the same class as \code{txt.file}.
}
\description{
This method calls \code{\link[koRpus:readability]{readability}} on all tagged text objects
inside the given \code{txt.file} object (using \code{lapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(path.package("tm.plugin.koRpus"), "tests", "testthat", "samples"),
  hierarchy=list(
    Topic=c(
      C3S="C3S",
      GEMA="GEMA"
    ),
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
myTexts <- readability(myCorpus)
}
}
