% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_21_kRp.corpus-class_show.R
\docType{methods}
\name{show,kRp.flatHier-method}
\alias{show,kRp.flatHier-method}
\title{Show methods for kRp.flatHier objects}
\usage{
\S4method{show}{kRp.flatHier}(object)
}
\arguments{
\item{object}{An object of class \code{kRp.flatHier}.}
}
\description{
Show methods for S4 objects of class \code{\link[tm.plugin.koRpus:kRp.flatHier-class]{kRp.flatHier}}.
}
