% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sourcesCorpus.R
\name{sourcesCorpus}
\alias{sourcesCorpus}
\title{Function to create kRp.sourcesCorpus objects from directory content}
\usage{
sourcesCorpus(path, sources, topic = "", ...)
}
\arguments{
\item{path}{A charcter vector with the path to the texts to analyse. Below this path,
      texts
must be ordered into subfolders named exactly like defined by \code{sources}.}

\item{sources}{A named character vector defining all sources to regard. The names of each entry will be used
internally for the sources.}

\item{topic}{A character string naming the topic covered.}

\item{...}{Additional options, passed through to \code{simpleCorpus}.}
}
\description{
Function to create kRp.sourcesCorpus objects from directory content
}
\examples{
\dontrun{
myTopic <- sourcesCorpus("~/data/foo", sources=c(bar="The Bar Magazine",
      baz="BAZ Monthly"))
}
}

