% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_06_kRp.corpus-class_summary.R
\docType{methods}
\name{summary,kRp.hierarchy-method}
\alias{summary,kRp.hierarchy-method}
\alias{corpusSummary}
\alias{corpusSummary,kRp.hierarchy-method}
\alias{corpusSummary,-methods}
\alias{corpusSummary<-}
\alias{corpusSummary<-,kRp.hierarchy-method}
\alias{corpusSummary<-,-methods}
\alias{corpusSummary<-,hierarchy-method}
\title{Apply summary() to all texts in kRp.hierarchy objects}
\usage{
\S4method{summary}{kRp.hierarchy}(object, missing = NA, ...)

corpusSummary(obj)

\S4method{corpusSummary}{kRp.hierarchy}(obj)

corpusSummary(obj) <- value

\S4method{corpusSummary}{kRp.hierarchy}(obj) <- value
}
\arguments{
\item{object}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{missing}{Character string to use for missing values.}

\item{...}{Used for internal processes.}

\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{value}{The new value to replace the current with.}
}
\value{
An object of the same class as \code{object}.
}
\description{
This method performs a \code{summary} call on all text objects
inside the given \code{object} object (using \code{lapply}). Contrary to what
other summary methods do, these methods always return the full object with
an updated \code{summary} slot.
}
\details{
The methods for nested object classes also recursively invoke the summary
methods for lower corpus objects.

The \code{summary} slot contains a data.frame with aggregated information of
all texts that the respective object level contains.
}
\examples{
\dontrun{
myTexts <- simpleCorpus(dir=file.path("/home","me","textCorpus"))
summary(myTexts)
}
}
