% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_06_kRp.corpus-class_summary.R,
%   R/02_method_06b_kRp.corpus-class_summary_old.R
\docType{methods}
\name{summary,kRp.flatHier-method}
\alias{summary,kRp.flatHier-method}
\alias{corpusSummary}
\alias{corpusSummary,kRp.flatHier-method}
\alias{corpusSummary,-methods}
\alias{corpusSummary<-}
\alias{corpusSummary<-,kRp.flatHier-method}
\alias{corpusSummary<-,-methods}
\alias{summary,kRp.hierarchy-method}
\alias{corpusSummary,kRp.hierarchy-method}
\alias{corpusSummary<-,kRp.hierarchy-method}
\alias{corpusSummary<-,hierarchy-method}
\title{Apply summary() to all texts in kRp.flatHier objects}
\usage{
\S4method{summary}{kRp.flatHier}(object, missing = NA, ...)

corpusSummary(obj)

\S4method{corpusSummary}{kRp.flatHier}(obj)

corpusSummary(obj) <- value

\S4method{corpusSummary}{kRp.flatHier}(obj) <- value

\S4method{summary}{kRp.hierarchy}(object, missing = NA, ...)

\S4method{corpusSummary}{kRp.hierarchy}(obj)

\S4method{corpusSummary}{kRp.hierarchy}(obj) <- value
}
\arguments{
\item{object}{An object of class \code{\link[tm.plugin.koRpus:kRp.flatHier-class]{kRp.flatHier}}.}

\item{missing}{Character string to use for missing values.}

\item{...}{Used for internal processes.}

\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.flatHier-class]{kRp.flatHier}}.}

\item{value}{The new value to replace the current with.}

\item{object}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{missing}{Character string to use for missing values.}

\item{...}{Used for internal processes.}
}
\value{
An object of the same class as \code{object}.

An object of the same class as \code{object}.
}
\description{
This method performs a \code{summary} call on all text objects inside the given
\code{object} object. Contrary to what other summary methods do, this method
always returns the full object with an updated \code{summary} slot.

This method performs a \code{summary} call on all text objects
inside the given \code{object} object (using \code{lapply}). Contrary to what
other summary methods do, these methods always return the full object with
an updated \code{summary} slot.
}
\details{
The \code{summary} slot contains a data.frame with aggregated information of
all texts that the respective object contains.

\code{corpusSummary} is a simple method to get or set the \code{summary} slot
in kRp.flatHier objects directly.

The methods for nested object classes also recursively invoke the summary
methods for lower corpus objects.

The \code{summary} slot contains a data.frame with aggregated information of
all texts that the respective object level contains.

\code{corpusSummary} is a simple method to get or set the \code{summary} slot
in kRp.hierarchy objects directly.
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
myCorpus <- readability(myCorpus, summary=FALSE)
corpusSummary(myCorpus)
# add summaries
myCorpus <- summary(myCorpus)
corpusSummary(myCorpus)
}
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
myCorpus <- readability(myCorpus, summary=FALSE)
corpusSummary(myCorpus)
# add summaries
myCorpus <- summary(myCorpus)
corpusSummary(myCorpus)
}
}
