% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_13_kRp.corpus-class_textTransform.R
\docType{methods}
\name{textTransform,kRp.hierarchy-method}
\alias{textTransform,kRp.hierarchy-method}
\title{Apply textTransform() to all texts in kRp.hierarchy objects}
\usage{
\S4method{textTransform}{kRp.hierarchy}(txt,
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{txt}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:textTransform]{textTransform}}.}
}
\value{
An object of the same class as \code{txt}.
}
\description{
This method calls \code{\link[koRpus:textTransform]{textTransform}} on all tagged text objects
inside the given \code{txt} object (using \code{lapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
# remove all punctuation
myCorpus <- textTransform(myCorpus, scheme="minor")
}
}
