% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topicCorpus.R
\name{topicCorpus}
\alias{topicCorpus}
\title{Function to create kRp.topicCorpus objects from directory content}
\usage{
topicCorpus(paths, sources, ...)
}
\arguments{
\item{paths}{A charcter vector with the path to the texts to analyse. Below this path,
      texts
must be ordered into subfolders named exactly like defined by \code{sources}.}

\item{sources}{A named character vector defining all sources to regard. The names of each entry will be used
internally for the sources.}

\item{...}{Additional options, passed through to \code{sourcesCorpus}.}
}
\description{
Function to create kRp.topicCorpus objects from directory content
}
\examples{
\dontrun{
myOrderedCorpus <- topicCorpus(
  paths=c(
    lottery="~/data/foo/lottery",
    waste="~/data/foo/waste"),
  sources=c(
    bar="The Bar Magazine",
    baz="BAZ Monthly")
)
}
}

