% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{corpusTagged}
\alias{corpusTagged}
\alias{tm.plugin.koRpus-deprecated}
\alias{corpusTTR}
\alias{corpusLevel}
\alias{tm.plugin.koRpus-defunct}
\alias{corpusCategory}
\alias{corpusID}
\alias{corpusPath}
\title{Deprecated functions and methods}
\usage{
corpusTagged(obj, ...)

corpusTTR(obj, ...)

corpusLevel(...)

corpusCategory(...)

corpusID(...)

corpusPath(...)
}
\arguments{
\item{obj}{No longer used.}

\item{...}{No longer used.}
}
\description{
These functions were used in earlier versions of the package but either
replaced or removed.
}
