% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_05_kRp.corpus-class_freq.analysis.R
\docType{methods}
\name{freq.analysis,kRp.corpus-method}
\alias{freq.analysis,kRp.corpus-method}
\alias{freq.analysis,kRp.sourcesCorpus-method}
\alias{freq.analysis,kRp.topicCorpus-method}
\title{Apply freq.analysis() to all texts in kRp.corpus objects}
\usage{
\S4method{freq.analysis}{kRp.corpus}(txt.file, ...)

\S4method{freq.analysis}{kRp.sourcesCorpus}(txt.file, ...)

\S4method{freq.analysis}{kRp.topicCorpus}(txt.file, ...)
}
\arguments{
\item{txt.file}{An object of class \code{\link[tm.plugin.koRpus]{kRp.corpus-class}},
\code{\link[tm.plugin.koRpus]{kRp.sourcesCorpus-class}} or
\code{\link[tm.plugin.koRpus]{kRp.topicCorpus-class}}.}

\item{...}{options to pass through to \code{\link[koRpus:freq.analysis]{freq.analysis}}.}
}
\value{
An object of the same class as \code{txt.file}.
}
\description{
This method calls \code{\link[koRpus:freq.analysis]{freq.analysis}} on all tagged text objects
inside the given \code{txt.file} object (using \code{lapply}).
}
\examples{
\dontrun{
myTexts <- simpleCorpus(dir=file.path("/home","me","textCorpus"))
myTexts <- freq.analysis(myTexts)
}
}

