% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_05_kRp.corpus-class_freq.analysis.R
\docType{methods}
\name{freq.analysis,kRp.hierarchy-method}
\alias{freq.analysis,kRp.hierarchy-method}
\title{Apply freq.analysis() to all texts in kRp.hierarchy objects}
\usage{
\S4method{freq.analysis}{kRp.hierarchy}(txt.file,
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{txt.file}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:freq.analysis]{freq.analysis}}.}
}
\value{
An object of the same class as \code{txt.file}.
}
\description{
This method calls \code{\link[koRpus:freq.analysis]{freq.analysis}} on all tagged text objects
inside the given \code{txt.file} object (using \code{lapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
# this will call read.corp.custom() recursively
myCorpus <- read.corp.custom(myCorpus)
myCorpus <- freq.analysis(myCorpus)
}
}
