% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_10_kRp.corpus-class_jumbleWords.R
\docType{methods}
\name{jumbleWords,kRp.corpus-method}
\alias{jumbleWords,kRp.corpus-method}
\title{Apply jumbleWords() to all texts in kRp.corpus objects}
\usage{
\S4method{jumbleWords}{kRp.corpus}(words,
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{words}{An object of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:jumbleWords]{jumbleWords}}.}
}
\value{
An object of the same class as \code{words}.
}
\description{
This method calls \code{\link[koRpus:jumbleWords]{jumbleWords}} on all tagged text objects
inside the given \code{words} object (using \code{mclapply}).
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
# remove all punctuation
myCorpus <- jumbleWords(myCorpus)
}
}
