% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_04_kRp.corpus.R
\docType{class}
\name{kRp.corpus,-class}
\alias{kRp.corpus,-class}
\alias{kRp_corpus}
\alias{kRp.corpus-class}
\title{S4 Class kRp.corpus}
\description{
Objects of this class can contain multiple texts simultaneously. It supports both the \code{tm} package's
\code{\link[tm]{Corpus}} class and \code{koRpus}' own object classes and stores them in separated slots.
}
\details{
Objects should be created using the \code{\link[tm.plugin.koRpus:simpleCorpus]{simpleCorpus}} function.
}
\section{Slots}{

\describe{
\item{\code{summary}}{A summary data frame for the full corpus.}

\item{\code{meta}}{A named list. Can be used to store meta information. Currently,
      no particular format is defined.}

\item{\code{raw}}{A list of objects of class \code{\link[tm]{Corpus}}.}

\item{\code{tagged}}{A list of objects of class \code{kRp.taggedText} (a class union for tagged text objects).}

\item{\code{hyphen}}{A list of objects of class \code{\link[sylly:kRp.hyphen-class]{kRp.hyphen}}.}

\item{\code{TTR}}{A list of objects of class \code{\link[koRpus:kRp.TTR-class]{kRp.TTR}}.}

\item{\code{readability}}{A list of objects of class \code{\link[koRpus:kRp.readability-class]{kRp.readability}}.}

\item{\code{freq}}{A list with two elements,
      \code{texts} and \code{corpus}. Both hold objects of class \code{\link[koRpus:kRp.corp.freq-class]{kRp.corp.freq}},
where \code{texts} is a list of these objects (one for each text),
      and \code{corpus} is a single object for the full corpus.}
}}

\note{
There is also \code{\link[tm.plugin.koRpus:kRp.corpus_get-methods]{getter and setter methods}} for objects of this class.
}
\section{Contructor function}{

Should you need to manually generate objects of this class (which should rarely be the case),
      the contructor function 
\code{kRp_corpus(...)} can be used instead of
\code{new("kRp.corpus", ...)}. Whenever possible, stick to
\code{\link[tm.plugin.koRpus:simpleCorpus]{simpleCorpus}}.
}

\keyword{classes}
