% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02_method_20_kRp.corpus_get_set_is.R
\docType{methods}
\name{corpusTagged}
\alias{corpusHyphen}
\alias{corpusHyphen,-methods}
\alias{corpusHyphen,kRp.corpus-method}
\alias{corpusHyphen<-}
\alias{corpusHyphen<-,-methods}
\alias{corpusHyphen<-,kRp.corpus-method}
\alias{corpusMeta}
\alias{corpusMeta,-methods}
\alias{corpusMeta,kRp.corpus-method}
\alias{corpusMeta<-}
\alias{corpusMeta<-,-methods}
\alias{corpusMeta<-,kRp.corpus-method}
\alias{corpusReadability}
\alias{corpusReadability,-methods}
\alias{corpusReadability,kRp.corpus-method}
\alias{corpusReadability<-}
\alias{corpusReadability<-,-methods}
\alias{corpusReadability<-,kRp.corpus-method}
\alias{corpusTTR}
\alias{corpusTTR,-methods}
\alias{corpusTTR,kRp.corpus-method}
\alias{corpusTTR<-}
\alias{corpusTTR<-,-methods}
\alias{corpusTTR<-,kRp.corpus-method}
\alias{corpusTagged}
\alias{corpusTagged,-methods}
\alias{corpusTagged,kRp.corpus-method}
\alias{corpusTagged<-}
\alias{corpusTagged<-,-methods}
\alias{corpusTagged<-,kRp.corpus-method}
\alias{corpusTm}
\alias{corpusTm,-methods}
\alias{corpusTm,kRp.corpus-method}
\alias{corpusTm<-}
\alias{corpusTm<-,-methods}
\alias{corpusTm<-,kRp.corpus-method}
\alias{is.corpus}
\title{Getter/setter methods for kRp.corpus objects}
\usage{
corpusTagged(obj)

\S4method{corpusTagged}{kRp.corpus}(obj)

corpusTagged(obj) <- value

\S4method{corpusTagged}{kRp.corpus}(obj) <- value

corpusReadability(obj)

\S4method{corpusReadability}{kRp.corpus}(obj)

corpusReadability(obj) <- value

\S4method{corpusReadability}{kRp.corpus}(obj) <- value

corpusTm(obj)

\S4method{corpusTm}{kRp.corpus}(obj)

corpusTm(obj) <- value

\S4method{corpusTm}{kRp.corpus}(obj) <- value

corpusMeta(obj)

\S4method{corpusMeta}{kRp.corpus}(obj)

corpusMeta(obj) <- value

\S4method{corpusMeta}{kRp.corpus}(obj) <- value

corpusHyphen(obj)

\S4method{corpusHyphen}{kRp.corpus}(obj)

corpusHyphen(obj) <- value

\S4method{corpusHyphen}{kRp.corpus}(obj) <- value

corpusTTR(obj)

\S4method{corpusTTR}{kRp.corpus}(obj)

corpusTTR(obj) <- value

\S4method{corpusTTR}{kRp.corpus}(obj) <- value

is.corpus(obj)
}
\arguments{
\item{obj}{An arbitrary \code{R} object.}

\item{value}{The new value to replace the current with.}

\item{value}{The new value to replace the current with.}

\item{value}{The new value to replace the current with.}

\item{value}{The new value to replace the current with.}

\item{value}{The new value to replace the current with.}

\item{value}{The new value to replace the current with.}
}
\description{
These methods should be used to get or set values of text objects
generated by functions like \code{\link[tm.plugin.koRpus:kRp.Corpus]{kRp.Corpus}}.
}
\details{
\itemize{
  \item{\code{corpusTagged()} }{returns the list of \code{kRp.tagged} objects.}
  \item{\code{corpusReadability()} }{returns the list of \code{kRp.readability} objects.}
  \item{\code{corpusTm()} }{returns the \code{VCorpus} object.}
  \item{\code{corpusMeta()} }{returns the list with meta information.}
  \item{\code{corpusHyphen()} }{returns the list of \code{kRp.hyphen} objects.}
  \item{\code{corpusTTR()} }{returns the list of \code{kRp.TTR} objects.}
}
}
\examples{
\dontrun{
corpusTagged(myCorpus)
corpusMeta(myCorpus)[["note"]] <- "an interesting read!"
}
}

