% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_20_kRp.corpus_get_set_is.R
\docType{methods}
\name{taggedText,kRp.flatHier-method}
\alias{taggedText,kRp.flatHier-method}
\alias{taggedText,-methods}
\alias{taggedText<-,kRp.flatHier-method}
\alias{taggedText<-,-methods}
\alias{describe,kRp.flatHier-method}
\alias{describe,-methods}
\alias{describe<-,kRp.flatHier-method}
\alias{describe<-,-methods}
\alias{language,kRp.flatHier-method}
\alias{language,-methods}
\alias{language<-,kRp.flatHier-method}
\alias{language<-,-methods}
\alias{corpusReadability}
\alias{corpusReadability,kRp.flatHier-method}
\alias{corpusReadability,-methods}
\alias{corpusReadability<-}
\alias{corpusReadability<-,kRp.flatHier-method}
\alias{corpusReadability<-,-methods}
\alias{corpusTm}
\alias{corpusTm,kRp.flatHier-method}
\alias{corpusTm,-methods}
\alias{corpusTm<-}
\alias{corpusTm<-,kRp.flatHier-method}
\alias{corpusTm<-,-methods}
\alias{corpusMeta}
\alias{corpusMeta,kRp.flatHier-method}
\alias{corpusMeta,-methods}
\alias{corpusMeta<-}
\alias{corpusMeta<-,kRp.flatHier-method}
\alias{corpusMeta<-,-methods}
\alias{corpusHyphen}
\alias{corpusHyphen,kRp.flatHier-method}
\alias{corpusHyphen,-methods}
\alias{corpusHyphen<-}
\alias{corpusHyphen<-,kRp.flatHier-method}
\alias{corpusHyphen<-,-methods}
\alias{corpusTTR}
\alias{corpusTTR,kRp.flatHier-method}
\alias{corpusTTR,-methods}
\alias{corpusTTR<-}
\alias{corpusTTR<-,kRp.flatHier-method}
\alias{corpusTTR<-,-methods}
\alias{corpusFreq}
\alias{corpusFreq,kRp.flatHier-method}
\alias{corpusFreq,-methods}
\alias{corpusFreq<-}
\alias{corpusFreq<-,kRp.flatHier-method}
\alias{corpusFreq<-,-methods}
\alias{corpusFiles}
\alias{corpusFiles,kRp.flatHier-method}
\alias{corpusFiles,-methods}
\alias{corpusFiles<-}
\alias{corpusFiles<-,kRp.flatHier-method}
\alias{corpusFiles<-,-methods}
\alias{diffText,kRp.flatHier-method}
\alias{diffText,-methods}
\alias{diffText<-,kRp.flatHier-method}
\alias{diffText<-,-methods}
\alias{originalText,kRp.flatHier-method}
\alias{originalText,-methods}
\alias{is.corpus}
\alias{[,kRp.flatHier,ANY,ANY,ANY-method}
\alias{[,-methods}
\alias{[<-,kRp.flatHier,ANY,ANY,ANY-method}
\alias{[<-,-methods}
\alias{[[,kRp.flatHier-method}
\alias{[[,-methods}
\alias{[[,kRp.flatHier,ANY-method}
\alias{[[<-,kRp.flatHier-method}
\alias{[[<-,-methods}
\alias{[[<-,kRp.flatHier,ANY,ANY-method}
\alias{tif_as_tokens_df}
\alias{tif_as_tokens_df,kRp.flatHier-method}
\alias{tif_as_tokens_df,-methods}
\alias{tif_as_tokens_df,hierarchy-method}
\alias{tif_as_corpus_df}
\alias{tif_as_corpus_df,kRp.flatHier-method}
\alias{tif_as_corpus_df,-methods}
\alias{tif_as_corpus_df,hierarchy-method}
\title{Getter/setter methods for kRp.flatHier objects}
\usage{
\S4method{taggedText}{kRp.flatHier}(obj)

\S4method{taggedText}{kRp.flatHier}(obj) <- value

\S4method{describe}{kRp.flatHier}(obj)

\S4method{describe}{kRp.flatHier}(obj) <- value

\S4method{language}{kRp.flatHier}(obj)

\S4method{language}{kRp.flatHier}(obj) <- value

corpusReadability(obj, ...)

\S4method{corpusReadability}{kRp.flatHier}(obj)

corpusReadability(obj) <- value

\S4method{corpusReadability}{kRp.flatHier}(obj) <- value

corpusTm(obj, id = NULL)

\S4method{corpusTm}{kRp.flatHier}(obj)

corpusTm(obj) <- value

\S4method{corpusTm}{kRp.flatHier}(obj) <- value

corpusMeta(obj, meta = NULL, fail = TRUE)

\S4method{corpusMeta}{kRp.flatHier}(obj, meta = NULL, fail = TRUE)

corpusMeta(obj, meta = NULL) <- value

\S4method{corpusMeta}{kRp.flatHier}(obj, meta = NULL) <- value

corpusHyphen(obj, ...)

\S4method{corpusHyphen}{kRp.flatHier}(obj)

corpusHyphen(obj) <- value

\S4method{corpusHyphen}{kRp.flatHier}(obj) <- value

corpusTTR(obj, ...)

\S4method{corpusTTR}{kRp.flatHier}(obj)

corpusTTR(obj) <- value

\S4method{corpusTTR}{kRp.flatHier}(obj) <- value

corpusFreq(obj, ...)

\S4method{corpusFreq}{kRp.flatHier}(obj)

corpusFreq(obj) <- value

\S4method{corpusFreq}{kRp.flatHier}(obj) <- value

corpusFiles(obj, paths = FALSE, ...)

\S4method{corpusFiles}{kRp.flatHier}(obj, paths = FALSE)

corpusFiles(obj) <- value

\S4method{corpusFiles}{kRp.flatHier}(obj) <- value

\S4method{diffText}{kRp.flatHier}(obj)

\S4method{diffText}{kRp.flatHier}(obj) <- value

\S4method{originalText}{kRp.flatHier}(obj)

is.corpus(obj)

\S4method{[}{kRp.flatHier,ANY,ANY,ANY}(x, i, j)

\S4method{[}{kRp.flatHier,ANY,ANY,ANY}(x, i, j) <- value

\S4method{[[}{kRp.flatHier}(x, i)

\S4method{[[}{kRp.flatHier}(x, i) <- value

tif_as_tokens_df(tokens)

\S4method{tif_as_tokens_df}{kRp.flatHier}(tokens)

tif_as_corpus_df(corpus)

\S4method{tif_as_corpus_df}{kRp.flatHier}(corpus)
}
\arguments{
\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.flatHier-class]{kRp.flatHier}}.}

\item{id}{Character value specifying the category ID of child objects you want to get.
If \code{NULL} then the child objects of the current top level are returned,
otherwise all children with the ID specified. Only interpreted if \code{level=NULL}
(you can't use both at the same time).}

\item{meta}{If not NULL, the \code{meta} list entry of the given name.}

\item{fail}{Logical,
      whether the method should fail with an error if \code{meta} was not found.
If set to \code{FALSE}, returns \code{invisible(NULL)} instead.}

\item{paths}{Logical,
      indicates for \code{corpusFiles()} whether full paths should be returned, or just the actual file name.}

\item{x}{See \code{obj}.}

\item{i}{Defines the row selector (\code{[}) or the name to match (\code{[[}) in the summary slot.}

\item{j}{Defines the column selector in the summary slot.}

\item{tokens}{An object of class \code{kRp.flatHier}.}

\item{corpus}{An object of class \code{kRp.flatHier}.}

\item{level}{Either the integer value or name (character string) of the target level. If \code{NULL} the top level is used.}
}
\description{
These methods should be used to get or set values of text objects
generated by functions like \code{\link[tm.plugin.koRpus:readCorpus]{readCorpus}}.
}
\details{
\itemize{
  \item{\code{taggedText()} }{returns the \code{TT.res} slot.}
  \item{\code{describe()} }{returns the \code{desc} slot.}
  \item{\code{corpusReadability()} }{returns the list of \code{kRp.readability} objects.}
  \item{\code{corpusTm()} }{returns the \code{VCorpus} object.}
  \item{\code{corpusMeta()} }{returns the list with meta information.}
  \item{\code{corpusHyphen()} }{returns the list of \code{kRp.hyphen} objects.}
  \item{\code{corpusTTR()} }{returns the list of \code{kRp.TTR} objects.}
  \item{\code{corpusFiles()} }{returns the character vector of file names of the object.}
  \item{\code{[}/\code{[[} }{can be used as a shortcut to index the results of \code{corpusSummary()}.}
  \item{\code{tif_as_corpus_df} }{returns the whole corpus in a single TIF[1] compliant
     data.frame.}
  \item{\code{tif_as_tokens_df} }{returns the \code{TT.res} slots of all texts in a single TIF[1] compliant
     data.frame, i.e., \code{doc_id} is not a factor but a character vector.}
}

Please note that the \code{<-} methods usually do not work on nested levels but only on single object nodes.
}
\examples{
\dontrun{
taggedText(myCorpus)
corpusMeta(myCorpus, "note") <- "an interesting read!"

# export object to TIF compliant data frame
myCorpus_df <- tif_as_corpus_df(myCorpus)
}
}
\references{
[1] Text Interchange Formats (\url{https://github.com/ropensci/tif})
}
