% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_class_02_kRp.sourcesCorpus.R
\docType{class}
\name{kRp.sourcesCorpus,-class}
\alias{kRp.sourcesCorpus,-class}
\alias{kRp.sourcesCorpus-class}
\title{S4 Class kRp.sourcesCorpus}
\description{
Objects of this class can contain multiple texts simultaneously. Adding to that,
      these texts are ordered
by their source in a number of slots. Each slot is a list of entries, one for each source.
}
\section{Slots}{

\describe{
\item{\code{summary}}{A summary data.frame for all sources combined.}

\item{\code{paths}}{A list of character strings with paths to all files}

\item{\code{sources}}{A list of objects of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}}}

\item{\code{files}}{A list of character strings with only the file names of all texts}

\item{\code{freq}}{An object of class \code{\link[koRpus]{kRp.corp.freq-class}},
      can contain word frequency
information on the full corpus if this object was analysed with
\code{\link[tm.plugin.koRpus:read.corp.custom]{read.corp.custom}}.}
}}
\keyword{classes}

