% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/02_method_04_kRp.corpus-class_read.corp.custom.R
\docType{methods}
\name{read.corp.custom,kRp.hierarchy-method}
\alias{read.corp.custom,kRp.hierarchy-method}
\title{Apply read.corp.custom() to all texts in kRp.hierarchy objects}
\usage{
\S4method{read.corp.custom}{kRp.hierarchy}(corpus,
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{corpus}{An object of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:read.corp.custom]{read.corp.custom}}.}
}
\value{
An object of the same class as \code{corpus}.
}
\description{
This method calls \code{\link[koRpus:read.corp.custom]{read.corp.custom}} on all tagged text objects
inside the given \code{corpus} object (using \code{lapply}).
}
\examples{
\dontrun{
myBasePath <- file.path("/home","me","textCorpus")
# analyse a single corpus
myTexts <- simpleCorpus(dir=myBasePath)
myTexts <- read.corp.custom(myTexts)

# you can also analyse multiple corpora from different sources simultaneously
# the following would assume that below 'myBasePath', there are subfolders
# named "Wikipedia" and Journal of Applied Geschwurbel"
mySources <- c(
   wp="Wikipedia",
   jg="Journal of Applied Geschwurbel"
 )
mySourcesTexts <- sourcesCorpus(
  path=myBasePath,
  sources=mySources
)
# this will also call read.corp.custom() recursively on
# the single text level
mySourcesTexts <- read.corp.custom(mySourcesTexts)

# and you might have guessed, you can also add a topic level
# with the following two vectors you would describe your
# data structure as two subfolders called "proc" and "gesw"
# below myBasePath, and two further subfolders named "Wikipedia"
# and Journal of Applied Geschwurbel" below each of the topic folders,
# containing the actual texts
myTopicPaths <- c(
  procrastination=file.path(myBasePath, "proc"),
  geschwurbel=file.path(myBasePath, "gesw")
)
mySources <- c(
   wp="Wikipedia",
   jg="Journal of Applied Geschwurbel"
 )
myTopicTexts <- topicCorpus(
  paths=myTopicPaths,
  sources=mySources
)
# this will also call read.corp.custom() recursively on
# the source and single text level
myTopicTexts <- read.corp.custom(myTopicTexts)

}
}
