% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/02_method_04_kRp.corpus-class_read.corp.custom.R
\docType{methods}
\name{read.corp.custom,kRp.flatHier-method}
\alias{read.corp.custom,kRp.flatHier-method}
\title{Apply read.corp.custom() to all texts in kRp.flatHier objects}
\usage{
\S4method{read.corp.custom}{kRp.flatHier}(corpus,
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{corpus}{An object of class \code{\link[tm.plugin.koRpus:kRp.flatHier-class]{kRp.flatHier}}.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.}

\item{...}{options to pass through to \code{\link[koRpus:read.corp.custom]{read.corp.custom}}.}
}
\value{
An object of the same class as \code{corpus}.
}
\description{
This method calls \code{\link[koRpus:read.corp.custom]{read.corp.custom}} on all tagged text objects
inside the given \code{corpus} object.
}
\examples{
\dontrun{
myCorpus <- readCorpus(
  dir=file.path(
    path.package("tm.plugin.koRpus"), "tests", "testthat", "samples", "C3S"
  ),
  hierarchy=list(
    Source=c(
      Wikipedia_alt="Wikipedia (alt)",
      Wikipedia_neu="Wikipedia (neu)"
    )
  )
)
# this will call read.corp.custom() recursively
myCorpus <- read.corp.custom(myCorpus)
}
}
