% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_21_kRp.corpus-class_show.R,
%   R/02_method_21b_kRp.corpus-class_show_old.R
\docType{methods}
\name{show,kRp.flatHier-method}
\alias{show,kRp.flatHier-method}
\alias{show,kRp.hierarchy-method}
\title{Show methods for kRp.flatHier objects}
\usage{
\S4method{show}{kRp.flatHier}(object)

\S4method{show}{kRp.hierarchy}(object)
}
\arguments{
\item{object}{An object of class \code{kRp.flatHier}.}

\item{object}{An object of class \code{kRp.hierarchy}.}
}
\description{
Show methods for S4 objects of class \code{\link[tm.plugin.koRpus:kRp.flatHier-class]{kRp.flatHier}}.

Show methods for S4 objects of class \code{\link[tm.plugin.koRpus:kRp.hierarchy-class]{kRp.hierarchy}}.
}
