% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_21_kRp.corpus-class_show.R
\docType{methods}
\name{show,kRp.corpus-method}
\alias{show,kRp.corpus-method}
\title{Show methods for kRp.corpus objects}
\usage{
\S4method{show}{kRp.corpus}(object)
}
\arguments{
\item{object}{An object of class \code{kRp.corpus}.}
}
\description{
Show methods for S4 objects of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}}.
}
