% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcesCorpus.R
\name{sourcesCorpus}
\alias{sourcesCorpus}
\title{Function to create kRp.sourcesCorpus objects from directory or object content}
\usage{
sourcesCorpus(path, sources, topic = "", format = "file",
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{path}{Usually a charcter vector with the path to the texts to analyse. Below this path,
      texts
must be ordered into subfolders named exactly like defined by \code{sources}. However,
      if \code{format="obj"},
you use this argument to provide the texts themselves as a list of named character vectors,
      where the name
must match the names of \code{sources} and each element of the vectors is a single text.}

\item{sources}{A named character vector defining all sources to regard. The names of each entry will be used
internally for the sources.}

\item{topic}{A character string naming the topic covered.}

\item{format}{Either "file" or "obj",
      depending on whether you want to scan files or analyze the text in a given object, like
a character vector. See \code{\link[tm.plugin.koRpus:simpleCorpus]{simpleCorpus}} for more information.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.
This value is passed through to simpleCorpus.}

\item{...}{Additional options, passed through to \code{simpleCorpus}.}
}
\description{
Internally,
      this is a wrapper for \code{\link[tm.plugin.koRpus:simpleCorpus]{simpleCorpus}} you can use
to analyze texts from more than one source.
}
\examples{
\dontrun{
myTopic <- sourcesCorpus("~/data/foo", sources=c(bar="The Bar Magazine",
      baz="BAZ Monthly"))

# providing text directly
myTopic <- sourcesCorpus(
  list(
    source1=c("the first text.", "the second text."),
    source2=c("the third text.", "the last text.")
  ),
  sources=c(
    source1="The Important",
    source2="Must Read Magazine"
  ),
  topic="short stories",
  format="obj"
)
}
}

