% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_method_06_kRp.corpus-class_summary.R
\docType{methods}
\name{summary,kRp.corpus-method}
\alias{summary,kRp.corpus-method}
\alias{summary}
\alias{summary,kRp.sourcesCorpus-method}
\alias{summary,kRp.topicCorpus-method}
\alias{corpusSummary}
\alias{corpusSummary,kRp.corpus-method}
\alias{corpusSummary,-methods}
\alias{corpusSummary,kRp.sourcesCorpus-method}
\alias{corpusSummary,-methods}
\alias{corpusSummary,kRp.topicCorpus-method}
\alias{corpusSummary,-methods}
\alias{corpusSummary<-}
\alias{corpusSummary<-,kRp.corpus-method}
\alias{corpusSummary<-,-methods}
\alias{corpusSummary<-,kRp.sourcesCorpus-method}
\alias{corpusSummary<-,-methods}
\alias{corpusSummary<-,kRp.topicCorpus-method}
\alias{corpusSummary<-,-methods}
\title{Apply summary() to all texts in kRp.corpus objects}
\usage{
\S4method{summary}{kRp.corpus}(object, missing = NA, ...)

\S4method{summary}{kRp.sourcesCorpus}(object, ...)

\S4method{summary}{kRp.topicCorpus}(object)

corpusSummary(obj)

\S4method{corpusSummary}{kRp.corpus}(obj)

\S4method{corpusSummary}{kRp.sourcesCorpus}(obj)

\S4method{corpusSummary}{kRp.topicCorpus}(obj)

corpusSummary(obj) <- value

\S4method{corpusSummary}{kRp.corpus}(obj) <- value

\S4method{corpusSummary}{kRp.sourcesCorpus}(obj) <- value

\S4method{corpusSummary}{kRp.topicCorpus}(obj) <- value
}
\arguments{
\item{object}{An object of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}},
\code{\link[tm.plugin.koRpus:kRp.sourcesCorpus-class]{kRp.sourcesCorpus}} or
\code{\link[tm.plugin.koRpus:kRp.topicCorpus-class]{kRp.topicCorpus}}.}

\item{missing}{Character string to use for missing values.}

\item{...}{Used for internal processes.}

\item{obj}{An object of class \code{\link[tm.plugin.koRpus:kRp.corpus-class]{kRp.corpus}},
\code{\link[tm.plugin.koRpus:kRp.sourcesCorpus-class]{kRp.sourcesCorpus}} or
\code{\link[tm.plugin.koRpus:kRp.topicCorpus-class]{kRp.topicCorpus}}.}

\item{value}{The new value to replace the current with.}
}
\value{
An object of the same class as \code{object}.
}
\description{
This method performs a \code{summary} call on all text objects
inside the given \code{object} object (using \code{lapply}). Contrary to what
other summary methods do, these methods always return the full object with
an updated \code{summary} slot.
}
\details{
The methods for nested object classes also recursively invoke the summary
methods for lower corpus object classes; i.e., if you call \code{summary}
on an object of class \code{kRp.topicCorpus}, the nested objects of class
\code{kRp.sourcesCorpus} will also be summarised, and that in turn causes
summaries on all nested corpora of class \code{kRp.corpus}.

The \code{summary} slot contains a data.frame with aggregated information of
all texts that the respective object level contains.
}
\examples{
\dontrun{
myTexts <- simpleCorpus(dir=file.path("/home","me","textCorpus"))
summary(myTexts)
}
}
