% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm.plugin.koRpus-package.R
\docType{package}
\name{tm.plugin.koRpus-package}
\alias{tm.plugin.koRpus-package}
\title{The tm.plugin.koRpus Package}
\description{
A Compatibility Plugin Package for 'tm' and 'koRpus'.
}
\details{
\tabular{ll}{
Package: \tab tm.plugin.koRpus\cr
Type: \tab Package\cr
Version: \tab 0.01-4\cr
Date: \tab 2016-07-12\cr
Depends: \tab R (>= 2.10.0),methods,parallel,koRpus (>= 0.07-1),tm\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://reaktanz.de/?c=hacking&s=koRpus\cr
}

This package provides classes and methods to enhance the ability to use the 'koRpus'
package together with the 'tm' package. It is in its early stages. To ask for help, report
bugs, suggest feature improvements, or discuss the global development of the package,
      please
subscribe to the koRpus-dev mailing list:
https://ml06.ispgateway.de/mailman/listinfo/korpus-dev_r.reaktanz.de
}
\author{
m.eik michalke
}
\keyword{package}

