% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm.plugin.koRpus-package.R
\docType{package}
\name{tm.plugin.koRpus-package}
\alias{tm.plugin.koRpus}
\alias{tm.plugin.koRpus-package}
\title{\packageTitle{tm.plugin.koRpus}}
\description{
\packageDescription{tm.plugin.koRpus}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab tm.plugin.koRpus\cr
Type: \tab Package\cr
Version: \tab 0.4-1\cr
Date: \tab 2020-12-17\cr
Depends: \tab R (>= 3.5.0),koRpus (>= 0.13-1),sylly (>= 0.1-6)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://reaktanz.de/?c=hacking&s=koRpus\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://reaktanz.de/?c=hacking&s=koRpus}
  \item Report bugs at \url{https://github.com/unDocUMeantIt/tm.plugin.koRpus/issues}
}

}
\author{
\packageAuthor{tm.plugin.koRpus}

Maintainer: \packageMaintainer{tm.plugin.koRpus}
}
