% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topicCorpus.R
\name{topicCorpus}
\alias{topicCorpus}
\title{Function to create kRp.topicCorpus objects from directory or object content}
\usage{
topicCorpus(paths, sources, format = "file",
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{paths}{A named list,
      usually with the paths to the texts to analyze. Each named list element
will define one topic. Below this path,
      texts must be ordered into subfolders named exactly like defined
by \code{sources}. However, if \code{format="obj"},
      you use this argument to provide the texts themselves
as a list of named lists of named character vectors,
      where the names must match the names of \code{sources}
and each element of the vectors is a single text. See the examples below.}

\item{sources}{A named character vector defining all sources to regard. The names of each entry will be used
internally for the sources.}

\item{format}{Either "file" or "obj",
      depending on whether you want to scan files or analyze the text in a given object, like
a character vector. See \code{\link[tm.plugin.koRpus:sourcesCorpus]{sourcesCorpus}} for more information.}

\item{mc.cores}{The number of cores to use for parallelization,
      see \code{\link[parallel:mclapply]{mclapply}}.
This value is passed through to simpleCorpus.}

\item{...}{Additional options, passed through to \code{sourcesCorpus}.}
}
\description{
Internally,
      this is a wrapper for \code{\link[tm.plugin.koRpus:sourcesCorpus]{sourcesCorpus}} you can use
to analyze texts both from more than one source, and on different topics.
}
\examples{
\dontrun{
myOrderedCorpus <- topicCorpus(
  paths=list(
    lottery="~/data/foo/lottery",
    waste="~/data/foo/waste"),
  sources=c(
    bar="The Bar Magazine",
    baz="BAZ Monthly")
)

# providing texts directly
myOrderedCorpus <- topicCorpus(
  paths=list(
    lottery=list(
      source1=c("the first text.", "the second text."),
      source2=c("the third text.", "the fourth text.")
    ),
    waste=list(
      source1=c("the fifth text.", "the sixth text."),
      source2=c("the seventh text.", "the last text.")
    )
  ),
  sources=c(
    source1="The Important",
    source2="Must Read Magazine"
  ),
  format="obj"
)
}
}

